# Conflicts:
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/common/SendMailCommandHandler.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/find/application/common/ApplicationFinder.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/after/DetailAfterRemandImpl.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/mail/CheckTranmissionImpl.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.infra/src/main/java/nts/uk/ctx/at/request/infra/repository/application/common/JpaApplicationRepository_New.java
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/000/b/kaf000.b.start.ts
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pub/src/main/java/nts/uk/ctx/bs/employee/pub/employee/SyEmployeePub.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pubimp/src/main/java/nts/uk/ctx/bs/employee/pubimp/contact/EmployeeContactPubImpl.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pubimp/src/main/java/nts/uk/ctx/bs/employee/pubimp/employee/SyEmployeePubImp.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.app/src/main/java/nts/uk/ctx/sys/env/app/command/mailnoticeset/company/dto/MailDestinationFunctionDto.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.dom/src/main/java/nts/uk/ctx/sys/env/dom/mailnoticeset/company/MailDestinationFunction.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.dom/src/main/java/nts/uk/ctx/sys/env/dom/mailnoticeset/company/MailDestinationFunctionGetMemento.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.dom/src/main/java/nts/uk/ctx/sys/env/dom/mailnoticeset/company/MailDestinationFunctionRepository.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.dom/src/main/java/nts/uk/ctx/sys/env/dom/mailnoticeset/employee/UseContactSettingRepository.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.infra/src/main/java/nts/uk/ctx/sys/env/infra/repository/mailnoticeset/company/JpaMailDestinationFunctionGetMemento.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.infra/src/main/java/nts/uk/ctx/sys/env/infra/repository/mailnoticeset/company/JpaMailDestinationFunctionRepository.java
#	nts.uk/uk.com/com.ctx/sys/env/nts.uk.ctx.sys.env.infra/src/main/java/nts/uk/ctx/sys/env/infra/repository/mailnoticeset/employee/JpaUseContactSettingRepository.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/add/AddSpLea14InfoCommandHandler.java
@@ -14,12 +14,13 @@ import nts.uk.shr.pereg.app.command.PeregAddCommandHandler;
 import nts.uk.shr.pereg.app.command.PeregAddCommandResult;
 
 @Stateless
-public class AddSpLea14InfoCommandHandler extends CommandHandlerWithResult<AddSpecialleave14informationCommand, PeregAddCommandResult>
-implements PeregAddCommandHandler<AddSpecialleave14informationCommand>{
+public class AddSpLea14InfoCommandHandler
+		extends CommandHandlerWithResult<AddSpecialleave14informationCommand, PeregAddCommandResult>
+		implements PeregAddCommandHandler<AddSpecialleave14informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler addSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00052";
@@ -32,12 +33,13 @@ implements PeregAddCommandHandler<AddSpecialleave14informationCommand>{
 
 	@Override
 	protected PeregAddCommandResult handle(CommandHandlerContext<AddSpecialleave14informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
+
 		String cid = AppContexts.user().companyId();
-		
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00052.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
+
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00052.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
 				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		return new PeregAddCommandResult(addSpLeaInfoCommandHandler.addHandler(domain));
 	}
