+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/remainingnumber/nursingcareleave/data/CareLeaveDataDto.java
@@ -13,5 +13,5 @@ public class CareLeaveDataDto {
 	private String sId;
 	
 	//使用日数
-	private int numOfUsedDay;
+	private double numOfUsedDay;
 }
